Broke geometry.shapes into more managable and cohesive submodules
